* commit 'b9b3aeea4f2134e5a7e60f97236b8cbfa95c3c29':
  Prevent NullPointerExceptions for Protobuf adapters
  Check null access
  Fixed filtering method contexts
  Removed useless stream()
  Removed usage of useless mapping methods
  Optimized model by using concurrent list instead of copies
  ExecutionContext model improvements
  Added license
  Fixed dashboard.vm ExecutionContext access
  Added event for finalizing the execution model

+++ core-interop/src/main/java/eu/tsystems/mms/tic/testframework/report/model/context/report/Report.java
@@ -80,10 +80,15 @@ public class Report {
         String relativeReportDirString = PropertyManager.getProperty(TesterraProperties.REPORTDIR, DEFAULT_REPORTDIR);
         File finalReportDirectory = new File(relativeReportDirString);
         try {
-            FileUtils.deleteDirectory(finalReportDirectory);
-            FileUtils.moveDirectory(REPORT_DIRECTORY, finalReportDirectory);
-            REPORT_DIRECTORY = finalReportDirectory;
-            LOGGER.info("Report written to " + finalReportDirectory.getAbsolutePath());
+            if (finalReportDirectory.exists()) {
+                FileUtils.deleteDirectory(finalReportDirectory);
+            }
+
+            if (REPORT_DIRECTORY.exists()) {
+                FileUtils.moveDirectory(REPORT_DIRECTORY, finalReportDirectory);
+                REPORT_DIRECTORY = finalReportDirectory;
+                LOGGER.info("Report written to " + finalReportDirectory.getAbsolutePath());
+            }
         } catch (IOException e) {
             throw new TesterraRuntimeException("Could not move report dir: " + e.getMessage(), e);
         }
